+++ index.html
@@ -1,53 +1,67 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>JakSim3 - 최고 평점 영화 컬렉션</title>
-    <link rel="stylesheet" href="./css/style.css">
-    <link rel="stylesheet" href="./css/reset.css">
-</head>
-<body>
+    <link rel="stylesheet" href="./css/style.css" />
+    <link rel="stylesheet" href="./css/reset.css" />
+  </head>
+  <body>
     <!-- header -->
     <header id="header">
-        <h1>
-            <a href="#">JakSim3 
-                <span class="intro">최고 평점 영화 컬렉션</span>
-            </a>
-        </h1>
+      <h1>
+        <a href="#"
+          >JakSim3
+          <span class="intro">최고 평점 영화 컬렉션</span>
+        </a>
+      </h1>
 
-        <!-- 언어 선택 -->
-        <div class="dropBox">
-            <button type="button" class="naBtn" id="nation">국가</button>
-            <div class="dropDown hidden">
-              <button type="button" id="nationBtn">한국</button>
-              <button type="button" id="nationBtn">미국</button>
-              <button type="button" id="nationBtn">중국</button>
-              <button type="button" id="nationBtn">일본</button>
-            </div>
+      <!-- 언어 선택 -->
+      <div class="dropBox">
+        <button type="button" class="naBtn" id="nation">국가</button>
+        <div class="dropDown hidden">
+          <button type="button" id="nationBtn">한국</button>
+          <button type="button" id="nationBtn">미국</button>
+          <button type="button" id="nationBtn">중국</button>
+          <button type="button" id="nationBtn">일본</button>
         </div>
+      </div>
     </header>
 
     <main id="container" class="main">
-        <!-- 영화 검색 -->
-        <form class="search" action="./index.html" method="POST" onsubmit="movieSearch()">
-            <label id="search">영화 검색 : </label>
-            <input type="text" id="searchInput" placeholder="영화제목을 입력해주세요^^" required />
-            <button type="submit" id="searchBtn" >검색</button>
-        </form>
+      <!-- 영화 검색 -->
+      <form
+        class="search"
+        action="./index.html"
+        method="POST"
+        onsubmit="movieSearch()"
+      >
+        <label id="search">영화 검색 : </label>
+        <input
+          type="text"
+          id="searchInput"
+          placeholder="영화제목을 입력해주세요^^"
+          required
+        />
+        <button type="submit" id="searchBtn">검색</button>
+      </form>
 
-        <!-- 카드 파트 -->
-        <section class="cardList"></section>
+      <!-- 카드 파트 -->
+
+      <section class="cardList" onclick="location.href='./sub.html'"></section>
+      
     </main>
 
     <footer id="footer">
-        <p>made by JakSim3Jo</p>
+      <p>made by JakSim3Jo</p>
     </footer>
 
-
     <!-- script part -->
-    <script src="./js/main.js"></script>
-    <script src="./js/movieApi.js"></script>
-</body>
+    <script src="./js/movieAPI.js"></script>
+    <script src="./js/movieSearch.js"></script>
+    <script src="./js/cardView.js"></script>
+  </body>
+  </body>
 </html>
