+++ js/datailView.js
@@ -1,0 +1,2 @@
+// 상세페이지 데이터 view
+function datailView() {}
