+++ js/review.js
@@ -1,0 +1,1 @@
+function review() {}
